+++ ProjectServer/JsonResponsePacketSerializer.h
@@ -129,7 +129,7 @@ struct PlayerResults
 	std::string username;
 	unsigned int correctAnswerCount;
 	unsigned int wrongAnswerCount;
-	unsigned int averageAnswerTime;
+	double averageAnswerTime;
 };
 
 struct GetGameResultsResponse
